requirements.txt
WTForms==2.2.1
joblib==0.14.1
numpy==1.17.3
Flask==1.1.2
Pillow==7.1.2
scikit_learn==0.23.1
gunicorn==20.0.4
